create event folder
